Add: publicrestaurant.html and publicmenu.html to protect menu pages

+++ finalproject.py
@@ -1,6 +1,6 @@
 from flask import Flask, render_template, request
 from flask import redirect, url_for, jsonify, flash
-from sqlalchemy import create_engine
+from sqlalchemy import create_engine, asc
 from sqlalchemy.orm import sessionmaker
 from database_setup import Base, Restaurant, MenuItem, User
 
@@ -183,8 +183,11 @@ def menuItemJSON(restaurant_id, menu_id):
 @app.route('/restaurant/')
 def showRestaurants():
     # This page will show all my restaurants
-    restaurants = session.query(Restaurant).all()
-    return render_template('restaurants.html', restaurants=restaurants)
+    restaurants = session.query(Restaurant).order_by(asc(Restaurant.name))
+    if 'username' not in login_session:
+        return render_template('publicrestaurants.html', restaurants=restaurants)
+    else:
+        return render_template('restaurants.html', restaurants=restaurants)
 
 
 @app.route('/restaurants/new', methods=['GET', 'POST'])
@@ -209,6 +212,8 @@ def editRestaurant(restaurant_id):
     if 'username' not in login_session:
         return redirect('/login')
     editRes = session.query(Restaurant).filter_by(id=restaurant_id).one()
+    if editRes.user_id != login_session['user_id']:
+        return "<script>function myFunction() {alert('You are not authorized to edit this restaurant. Please create your own restaurant in order to edit.');}</script><body onload='myFunction()''>"
     if request.method == 'POST':
         if request.form['name']:
             editRes.name = request.form['name']
@@ -226,6 +231,8 @@ def deleteRestaurant(restaurant_id):
     if 'username' not in login_session:
         return redirect('/login')
     deleteRes = session.query(Restaurant).filter_by(id=restaurant_id).one()
+    if deleteRes.user_id != login_session['user_id']:
+        return "<script>function myFunction() {alert('You are not authorized to delete this restaurant. Please create your own restaurant in order to delete.');}</script><body onload='myFunction()''>"
     if request.method == 'POST':
         session.delete(deleteRes)
         session.commit()
@@ -240,8 +247,14 @@ def deleteRestaurant(restaurant_id):
 def showMenu(restaurant_id):
     # This page is the menu for restaurant
     restaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
-    items = session.query(MenuItem).filter_by(restaurant_id=restaurant.id)
-    return render_template('menu.html', restaurant=restaurant, items=items)
+    creator = getUserInfo(restaurant.user_id)
+    items = session.query(MenuItem).filter_by(restaurant_id=restaurant.id).all()
+    if 'username' not in login_session or creator.id != login_session['user_id']:
+        return render_template('publicmenu.html', items=items,
+                               restaurant=restaurant, creator=creator)
+    else:
+        return render_template('menu.html', items=items,
+                               restaurant=restaurant, creator=creator)
 
 
 @app.route('/restaurants/<int:restaurant_id>/menu/new/',
@@ -250,11 +263,14 @@ def newMenuItem(restaurant_id):
     # This page is for making a new menu item for restaurant
     if 'username' not in login_session:
         return redirect('/login')
+    restaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
+    if login_session['user_id'] != restaurant.user_id:
+        return "<script>function myFunction() {alert('You are not authorized to add menu items to this restaurant. Please create your own restaurant in order to add items.');}</script><body onload='myFunction()''>"
     if request.method == 'POST':
         newItem = MenuItem(
             name=request.form['name'], price=request.form['price'],
             description=request.form['description'],
-            restaurant_id=restaurant_id, user_id=login_session['user_id'])
+            restaurant_id=restaurant_id, user_id=restaurant.user_id)
         session.add(newItem)
         session.commit()
         flash(str(newItem.name) + " has been created!")
@@ -270,11 +286,18 @@ def editMenuItem(restaurant_id, menu_id):
     if 'username' not in login_session:
         return redirect('/login')
     editItem = session.query(MenuItem).filter_by(id=menu_id).one()
+    restaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
+    if login_session['user_id'] != restaurant.user_id:
+        return "<script>function myFunction() {alert('You are not authorized to edit menu items to this restaurant. Please create your own restaurant in order to edit items.');}</script><body onload='myFunction()''>"
     if request.method == 'POST':
-        if request.form['name'] and request.form['price']:
+        if request.form['name']:
             editItem.name = request.form['name']
-            editItem.price = request.form['price']
+        if request.form['description']:
             editItem.description = request.form['description']
+        if request.form['price']:
+            editItem.price = request.form['price']
+        #if request.form['course']:
+         #   editItem.course = request.form['course']
         session.add(editItem)
         session.commit()
         flash(str(editItem.name) + " has been edited!")
@@ -292,6 +315,9 @@ def deleteMenuItem(restaurant_id, menu_id):
     if 'username' not in login_session:
         return redirect('/login')
     deleteItem = session.query(MenuItem).filter_by(id=menu_id).one()
+    restaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
+    if login_session['user_id'] != restaurant.user_id:
+        return "<script>function myFunction() {alert('You are not authorized to delete menu items to this restaurant. Please create your own restaurant in order to delete items.');}</script><body onload='myFunction()''>"
     if request.method == 'POST':
         session.delete(deleteItem)
         session.commit()
@@ -301,6 +327,7 @@ def deleteMenuItem(restaurant_id, menu_id):
         return render_template('deleteMenuItem.html', item=deleteItem)
 
 
+# User Helper Functions
 def getUserID(email):
     try:
         user = session.query(User).filter_by(email=email).one()
